requirements.txt
click
attrs
rich
